requirements.txt
crewai==0.11.0
python-dotenv==1.0.0
flask==3.0.0
openai==1.0.0
requests==2.31.0
pydantic==2.0.0
gunicorn==21.2.0
werkzeug==3.0.1
slack-bolt>=1.18.0
slack-sdk>=3.26.0 